requirements.txt
transformers>=4.38.0
datasets>=2.15.0
trl>=0.8.0
peft>=0.9.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.42.0
wandb
numpy
tqdm
tenacity
scipy
matplotlib
seaborn
pandas
python-dotenv
huggingface_hub
ipywidgets

